requirements.txt
fastapi==0.79.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
numpy==1.22.0
Pillow==9.0.1
# opencv-contrib-python-headless
# tensorflow-macos
# tensorflow
opencv-python
pandas==1.3.5
mediapipe
python-multipart
protobuf==3.20.*